fixing dublicate
