+++ src/Fluent.Calculations.Primitives.Tests/Expressions/ReflectionProviderTests.cs
@@ -1,17 +1,29 @@
-﻿using Fluent.Calculations.Primitives.Expressions.Capture;
+﻿using Fluent.Calculations.Primitives.BaseTypes;
+using Fluent.Calculations.Primitives.Expressions.Capture;
+using FluentAssertions;
+using System.Linq.Expressions;
 
 namespace Fluent.Calculations.Primitives.Tests.Expressions
 {
     public class ReflectionProviderTests
     {
-        private ReflectionProvider reflectionProvider = new ReflectionProvider();
+        ReflectionProvider memberAccessReflectionProvider = new ReflectionProvider();
 
         [Fact]
-        public void Test()
+        public void HasValue_ReturnsResult()
         {
+            Number testValue = Number.Of(1, "TEST-VALUE");
+            Expression<Func<Number>> testExpression = () => testValue;
+            IValue result = memberAccessReflectionProvider.GetValue(testExpression.Body);
+            result.Name.Should().Be(testValue.Name);
+        }
 
-
-
+        [Fact]
+        public void NullValue_Throws()
+        {
+            Number testValue = null;
+            Expression<Func<Number>> testExpression = () => testValue;
+            Assert.Throws<NullExpressionResultException>(() => memberAccessReflectionProvider.GetValue(testExpression.Body));
         }
     }
 }
